backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
trimesh==4.0.5
numpy-stl==3.0.1
numpy==1.24.3
scipy==1.11.4
meshio==5.3.4
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0

# For STEP/IGES support (optional, requires complex installation)
# pythonOCC-core==7.7.2

# For additional mesh processing
open3d==0.18.0
pymeshlab==2022.2.post4